# docs/swagger/swagger.json
{
    "openapi": "3.0.3",
    "info": {
      "title": "SmartSavings API",
      "description": "Formal specification of endpoints to be used in the SmartSavings Platform",
      "version": "1.0.0"
    },
    "servers": [
      {
        "url": "http://localhost:3000"
      }
    ],
...
    "paths": {
...
      "/user": {
        "post": {
          "tags": [
            "user"
          ],
          "summary": "Create user",
          "description": "Create a user account",
          "operationId": "createUser",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "successful operation",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ApiResponse"
                  }
                }
              }
            }
          }
        },
        "get": {
          "tags": [
            "user"
          ],
          "summary": "Get user by user id",
          "description": "This can only be done by the logged in user.",
          "operationId": "getUserById",
          "responses": {
            "200": {
              "description": "successful operation",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        }
      },
      "/user/login": {
        "post": {
          "tags": [
            "user"
          ],
          "summary": "Logs user into the system",
          "description": "Allows a user to log in",
          "operationId": "loginUser",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginUser"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "successful operation",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ApiResponse"
                  }
                }
              }
            },
            "400": {
              "description": "Invalid username/password supplied"
            }
          }
        }
      },
      "/user/refreshToken": {
        "post": {
          "tags": [
            "user"
          ],
          "summary": "Obtain an accessToken by refreshToken",
          "description": "Use RefreshToken",
          "operationId": "refreshToken",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshToken"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "successful operation",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ApiResponse"
                  }
                }
              }
            },
            "400": {
              "description": "Invalid username/password supplied"
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
        "User": {
          "type": "object",
          "properties": {
            "firstName": {
              "type": "string",
              "example": "User Name"
            },
            "lastName": {
              "type": "string",
              "example": "User Surname"
            },
            "dateBirth": {
              "type": "string",
              "example": "1997-01-30",
              "format": "date",
              "pattern": "YYYY-MM-DD"
            },
            "objetive": {
              "type": "string",
              "example": "Personal Objetive"
            },
            "email": {
              "type": "string",
              "example": "user@email.com"
            },
            "repeatEmail": {
              "type": "string",
              "example": "user@email.com"
            },
            "password": {
              "type": "string",
              "example": "12345@Aa"
            },
            "repeatPassword": {
              "type": "string",
              "example": "12345@Aa"
            }
          }
        },
        "LoginUser": {
          "type": "object",
          "properties": {
            "email": {
              "type": "string",
              "example": "test@test.com"
            },
            "password": {
              "type": "string",
              "example": "Aabb@1"
            }
          }
        },
        "RefreshToken": {
          "type": "object",
          "properties": {
            "refreshToken": {
              "type": "string",
              "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpYXQiOjE3MDM1MjczNzcsImV4cCI6MTcwMzYxMzc3Nywic3ViIjoidGVzdEB0ZXN0LmNvbSJ9.s9rDV-zVSLllLje7GjtAy6Sl2nV8SJ9DP4IYnEvDEbc"
            }
          }
        },
...
        "ApiResponse": {
          "type": "object",
          "properties": {
            "code": {
              "type": "integer",
              "format": "int32"
            },
            "type": {
              "type": "string"
            },
            "message": {
              "type": "string"
            }
          },
          "xml": {
            "name": "##default"
          }
        }
      },
      "securitySchemes": {
        "api_key": {
          "type": "apiKey",
          "name": "api_key",
          "in": "header"
        }
      }
    }
  }